Fixes Contact Modification Bug
PresenceContact was being constructed incorrectly in
some cases leading to contacts not being updated
in the EAB database and never getting capability
information.

This change fixes the bug and removes the existing
rows in the database. It then repopulates the DB
with correct cached numbers.

I have also dialed back the verbosity of the logging
in general and hid PII.

Bug: 37691655
Test: Manual
Change-Id: I74e45eb3bcf47c3e03651421479f59aeb6762571