Do not start init until ACTION_UCE_SERVICE_UP
We were trying to get the UCE service before it was sending
the correct intent, which would sometimes cause it to lock
up (in certain race conditions). Remove init when starting
RCS service as well as when IMS is connected.

The UCE_SERVICE_UP intent is sticky, so there is no problem
restarting the RcsService if it crashes.

Bug: 119309891
Test: Manual
Change-Id: I39eeb5ca497408893b8a3e4e5b5644cfcecadf35