Remove usage of deprecated Time class
Remove usage of deprecated Time class for formatting a date/time and
replace it with the internal
TimeMigrationUtils.formatMillisWithFixedFormat(long) which does
something similar but which won't break in 2038.

The output produced is slightly different.

The format:
"%m-%d %H:%M:%S"
becomes:
"%Y-%m-%d %H:%M:%S"

This may be an issue if the toString() is parsed for any
reason but it looks like it's used for debug / human consumption.

Bug: 16550209
Test: build / treehugger only
Change-Id: I5475777d0ea2709326f4d38a1c80407d63ebbfee
Merged-In: I5475777d0ea2709326f4d38a1c80407d63ebbfee