[RCS] cap: Fix to register ProvisioningChangedCallback
When device boot up, CapabilityPolling gets the defaultSubId by getDefaultSubsciptionId() in PresenceSetting.java.
PollingService also gets the defaultSubId by same function.
DEFAULT_SUBSCRIPTION_CHANGED intent is sent by SubscriptionManager, PollingService check whether CapabilityPolling is initialized or not and if it is, it calls handleDefaultSubscriptionChanged with defaultSubId.
Here are two problems.
1) If CapabilityPolling is not initialized, handleDefaultSubscriptionChanged will never be called.
2) Even if it is called, both defaultSubIds, one is defined at CapabilityPolling and the other is parameter of handleDefaultSubscriptionChanged set by PollingService, are same so it returns without registering ProvisioningChangedCallback.

This Patch is for fixing this problem.

Signed-off-by: s_h.jeon <s_h.jeon@samsung.com>
Change-Id: I0980f993d7ba5f9011796608467a1a7895194c66